extract more from js_ast
