+++ cmake/sources/ZigSources.txt
@@ -11,10 +11,17 @@ src/analytics/analytics_schema.zig
 src/analytics/analytics_thread.zig
 src/api/schema.zig
 src/ast/base.zig
+src/ast/BundledAst.zig
 src/ast/E.zig
 src/ast/Expr.zig
+src/ast/G.zig
 src/ast/Macro.zig
+src/ast/Op.zig
 src/ast/S.zig
+src/ast/Scope.zig
+src/ast/ServerComponentBoundary.zig
+src/ast/Stmt.zig
+src/ast/Symbol.zig
 src/async/posix_event_loop.zig
 src/async/stub_event_loop.zig
 src/async/windows_event_loop.zig
